add `NamedScoreComputation` in order to store names of score comps
I really don't like the way this is so verbose and clunky.
however it seems like the most expedient way to get to where we want to
be. Holochain doesn't make it easy to load multiple entries at once.
oh well.

it would be easier if we could store the Name in side of the
InterchangeEntry, but then we'd have to modify the IE struct.
I suppose we could parameterize it by a generic and have a field of that
type... 🤔