rework zomeApi into hcClient, modify backend for bridge calls
now encountering 2 different errors non-deterministically (it seems).

one is:

```
"Wasm error while working with Ribosome: Host(\"ZomeError: Zome not found: Zome 'hub_main' not found\")"
```

which is comprehensibel but feels wrong - I believe the hub dna/cell is
installed and activated, and `hub_main` is the correct name.

\...

the other seemed to maybe be the `todo!()` in our match of (bridge) `call`
failing and causing an `unreachable` error with a gnarly stacktrace.