use `compose_entry_hash_path` util fn; demo works
we seemed to have issues with the Display-ing of EntryHashes - sometimes
they would come out `0x...` and others `uhCE...`. this meant that paths
created with one wouldn't (ofc) be retrieved from the other.

`compose_entry_hash_path` gets the inner byte-vec of an EntryHash and
base64-encodes it, then appends that to the provided path_string.

this means if all widget & sensemaker code uses this approach, we'll
have agreement between gets/sets. base64 encoding seems
reasonable-enough for our purposes.